add error handling and dynamic query params

+++ src/main/java/com/cebrailerdogan/cvrservice/dto/CvrRegistryResponseDto.java
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import java.util.List;
 
 public record CvrRegistryResponseDto(
-        int vat,
+        long vat,
         String name,
         String address,
         @JsonProperty("zipcode")
